- Moved loader component to 'sanda' directory
