+++ src/components/sanda/Loader.vue
